refactor: extract cron job logic into testable functions
- Extract runDailyScan(client) from llmSummaryJob and export it
- Move dailyStats cron body into utils/dailyStatsRunner.js as runDailyStats(client)
- Slim jobs/dailyStats.js and jobs/llmSummaryJob.js to pure scheduling
- Update scripts/runSummary.js to use runDailyScan for consistent logging
- Add unit tests for llmSummaryJob and dailyStatsRunner
